checkpointing; problems with assemble_sample
