+++ docs/api.rst
@@ -2,57 +2,6 @@
 API Reference
 #############
 
-********
-Commands
-********
-
-Model fit and analysis
-======================
-
-compile_model
--------------
-.. autofunction:: cmdstanpy.compile_model
-
-sample
-------
-.. autofunction:: cmdstanpy.sample
-
-get_drawset
------------
-.. autofunction:: cmdstanpy.get_drawset
-
-save_csvfiles
--------------
-.. autofunction:: cmdstanpy.save_csvfiles
-
-summary
--------
-.. autofunction:: cmdstanpy.summary
-
-diagnose
---------
-.. autofunction:: cmdstanpy.diagnose
-
-Utility functions
-=================
-
-set_cmdstan_path
-----------------
-.. autofunction:: cmdstanpy.set_cmdstan_path
-
-cmdstan_path
-------------
-.. autofunction:: cmdstanpy.cmdstan_path
-
-jsondump
---------
-.. autofunction:: cmdstanpy.jsondump
-
-rdump
---------
-.. autofunction:: cmdstanpy.rdump
-
-
 *******
 Classes
 *******
@@ -64,8 +13,8 @@ Model
 .. autoclass:: cmdstanpy.Model
    :members:
 
-RunSet
-======
+StanFit
+=======
 
-.. autoclass:: cmdstanpy.RunSet
+.. autoclass:: cmdstanpy.StanFit
    :members:
